requirements.txt
flask==3.0
flask-login==0.6.3
flask-migrate==4.0.5
flask-sqlalchemy==3.1.1
flask-admin==1.6.1
flask-bcrypt==1.0.1
pytz==2023.3.post1
gunicorn==21.2.0
loguru==0.7.2
python-dotenv==1.0.0
flask-httpauth==4.8.0
flask-jwt-extended==4.5.3
requests==2.31.0
python-slugify==8.0.1
psycopg2-binary==2.9.9
flask-security-too==5.3.2
flask-apispec==0.11.4
flask-marshmallow==0.15.0
flask-restful==0.3.10
marshmallow-sqlalchemy==0.29.0
flasgger==0.9.7.1
jinja2==3.1.2
pytest==8.1.1
fake-useragent==1.5.1